api-viewer: use our extjs package instead of the cloudflare cdn
The loading of an external hosted ExtJS is from a time where we did
not have a extra ExtJS package as it still was in the pve-manager
one, which was not desired to be installed on a non-pmg machine.
Add a dependency on libjs-extjs and a link from api-viewer/extjs to
the installation directory of the ExtJS package, this allows to reuse
ExtJS on PMG node hosted docs and on the hosted docs from proxmox.com
at the same time.

Adapted from the Proxmox VE version[0] of this change.

[0]: https://git.proxmox.com/?p=pve-docs.git;a=commitdiff;h=e56030c4d7c3bd0de1ebc109d528fb6b7bef05a3

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>