fix lintian warning in pmg-docs-apache
the apache/conf-available snippet for pmg-docs-apache should be named like the
package according to lintian (non-standard-apache2-configuration-name).

move the configfile via dpkg-maintscript-helper(1) (see the manpage for
the explaination of the prior version of 6.0-3~) and dh_installdeb(1).

Tested by upgrading pmg-docs-apache to a version with this commit, while having
modifications in '/etc/apache2/conf-available/pmg-docs.conf'

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>