mail-filter: include 'Rule-Based' in heading
One of the core features of PMG is the rule-system, which is documented in
'pmg-mail-filter.adoc' under the (correct) heading of 'Mail Filter'

In quite a few other places [0,1] where the features of PMG are highlighted we
refer to the (Object Oriented) Rule System.

When skimming through the documentation and quickly glancing at links the
connection 'Rule System' <-> 'Mail Filter' is not directly apparent and could
cause confusion.

This is addressed by adding 'Rule-Based' as prefix to the chapter title.

[0] https://www.proxmox.com/en/proxmox-mail-gateway/features
[1] https://pmg.proxmox.com/pmg-docs/pmg-admin-guide.html#_features

Reported-by: Daniela Häsler <daniela@proxmox.com>
Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>