add big attribute to thumbnails
allows one to easily define bigger thumbnails, i.e., those where we
have to big images to show the full on a page, but still want to give
the user a good sized preview.

As PDF renders this already pretty nice by default we only have to
adapt the html asciidoc config. We do that by using asciidocs
"Conditional Attribute References"[0] and add a "big" attribute.

It can be used as:
> [thumbnail="pmg-gui-network-config.png", big=1]

NOTE: the value of big is ignored, it just matters if it's defined.
While it could be possible to add logic to handle big=0, it it's just
extra work without real use, just omit it if not wanted.

[0]: https://www.methods.co.nz/asciidoc/chunked/ch30.html

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>