clarify tracking center docs for before-queue
In the after-queue case we get 2 QEntrys from the pmg-log-tracker when
the mail is accepted. One serves as a relay and provides the delivery
status. In the before-queue case we only have 1 QEntry and no extra delivery
status. This results in before-queue accepted mails being shown as
'queued/accepted' instead of 'accepted/<status>'.

As the tracking center shows 'queued/accepted', mention this combination
in the docs and clarify that it only happens in the before-queue
filtering case.

Signed-off-by: Mira Limbeck <m.limbeck@proxmox.com>