asciidoc-pmg: ignore link targets for non-manpages
(This is taken from pve-docs commit
7de8bba68de1875aca04ebd1cd941a2992c24753)

To allow linking from to a chapter/section not included in a manpage
allow the manpage link resolver to just return text in a case the
link target text is in fact no manpage.

If the link is a valid one in general will be checked in a lot of
other places, so here we won't run into a regression where a wrong,
non-existing, link does not get detected by the build system.

Signed-off-by: Mira Limbeck <m.limbeck@proxmox.com>
Originally-by: Thomas Lamprecht <t.lamprecht@proxmox.com>